Refactor: move 'blener_verification_order' helper functions to 'utils' module and adapt tests
