+++ concent_api/verifier/tests/test_integration_verification.py
@@ -1,13 +1,11 @@
-from subprocess import SubprocessError
-from zipfile import BadZipFile
-import io
 import mock
 
-from django.conf import settings
 from django.test import override_settings
+from numpy.core.multiarray import ndarray
 
 from conductor.models import BlenderSubtaskDefinition
 from core.constants import VerificationResult
+from core.tasks import verification_result
 from core.tests.utils import ConcentIntegrationTestCase
 from common.constants import ErrorCode
 from common.helpers import get_storage_result_file_path
@@ -71,93 +69,73 @@ class VerifierVerificationIntegrationTest(ConcentIntegrationTestCase):
             )
         )
 
-    def test_that_blender_verification_order_should_perform_full_verification_with_match_result(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks,\
-            mock.patch('verifier.tasks.unpack_archive', autospec=True) as mock_unpack_archive,\
-            mock.patch('core.tasks.verification_result.delay', autospec=True) as mock_verification_result,\
-            mock.patch('verifier.tasks.run_blender', mock_run_blender),\
-            mock.patch('verifier.tasks.upload_file_to_storage_cluster', autospec=True) as mock_upload_file, \
-            mock.patch('builtins.open', autospec=True, side_effect=[io.BytesIO(b'test'), io.BytesIO(b'test')]), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.compare_ssim', return_value=1.0) as mock_compare_ssim, \
-            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True), \
-            mock.patch('verifier.tasks.delete_file', autospec=True) as mock_delete_file, \
-            mock.patch('verifier.tasks.import_cv2', return_value=self._prepare_cv2_mock()) as mock_import_cv2:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
+        self.subtask_id = self.compute_task_def['subtask_id']
+        self.scene_file = self.compute_task_def['extra_data']['scene_file']
+        self.output_format = self.compute_task_def['extra_data']['output_format']
 
-        self.assertEqual(mock_clean_directory.call_count, 1)
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 4)
-        self.assertEqual(mock_delete_file.call_count, 2)
-        self.assertEqual(mock_upload_file.call_count, 1)
-        mock_compare_ssim.assert_called_once()
-        mock_import_cv2.assert_called_once()
-        mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
+        self.mock_image1 = mock.create_autospec(spec=ndarray, spec_set=True)
+        self.mock_image2 = mock.create_autospec(spec=ndarray, spec_set=True)
+        self.loaded_images = (self.mock_image1, self.mock_image2)
+        self.mock_verification_result = mock.create_autospec(spec=verification_result, spec_set=True)
+
+    def test_that_blender_verification_order_should_perform_full_verification_with_match_result(self):
+        with mock.patch('verifier.tasks.download_archives_from_storage', autospec=True) as mock_download_archives_from_storage,\
+            mock.patch('verifier.tasks.validate_downloaded_archives', autospec=True) as mock_validate_downloaded_archives, \
+            mock.patch('verifier.tasks.unpack_archives', autospec=True) as mock_unpack_archives, \
+            mock.patch('verifier.tasks.render_image', autospec=True) as mock_render_image, \
+            mock.patch('verifier.tasks.delete_source_files', autospec=True) as mock_delete_source_files, \
+            mock.patch('verifier.tasks.try_to_upload_blender_output_file', autospec=True) as mock_try_to_upload_file, \
+            mock.patch('verifier.tasks.load_images', autospec=True, return_value=self.loaded_images) as mock_load_images, \
+            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True) as mock_are_image_sizes_and_color_channels_equal, \
+            mock.patch('verifier.tasks.compare_images', side_effect=self._compare_images_positive) as mock_compare_images:  # noqa: E125
+
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
+        self.assertEqual(mock_validate_downloaded_archives.call_count, 1)
+        self.assertEqual(mock_unpack_archives.call_count, 1)
+        mock_render_image.assert_called_once_with(1, self.output_format, self.scene_file, self.subtask_id)
+        self.assertEqual(mock_delete_source_files.call_count, 1)
+        self.assertEqual(mock_try_to_upload_file.call_count, 1)
+        self.assertEqual(mock_load_images.call_count, 1)
+        mock_are_image_sizes_and_color_channels_equal.assert_called_once_with(*self.loaded_images)
+        mock_compare_images.assert_called_once_with(self.mock_image1, self.mock_image2, self.subtask_id)
+        self.mock_verification_result.assert_called_once_with(
+            self.subtask_id,
             VerificationResult.MATCH.name,
         )
 
     def test_that_blender_verification_order_should_perform_full_verification_with_mismatch_result(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks,\
-            mock.patch('verifier.tasks.unpack_archive', autospec=True) as mock_unpack_archive, \
-            mock.patch('verifier.tasks.upload_file_to_storage_cluster', autospec=True) as mock_upload_file_to_storage_cluster, \
-            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result,\
-            mock.patch('verifier.tasks.run_blender', mock_run_blender), \
-            mock.patch('builtins.open', autospec=True, side_effect=[io.BytesIO(b'test'), io.BytesIO(b'test')]), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', autospec=True, return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.compare_ssim', return_value=(settings.VERIFIER_MIN_SSIM - 0.1)) as mock_compare_ssim, \
-            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True), \
-            mock.patch('verifier.tasks.delete_file', autospec=True) as mock_delete_file, \
-            mock.patch('verifier.tasks.import_cv2', return_value=self._prepare_cv2_mock()) as mock_import_cv2:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
+        with mock.patch('verifier.tasks.download_archives_from_storage', autospec=True) as mock_download_archives_from_storage, \
+            mock.patch('verifier.tasks.validate_downloaded_archives', autospec=True) as mock_validate_downloaded_archives, \
+            mock.patch('verifier.tasks.unpack_archives', autospec=True) as mock_unpack_archives, \
+            mock.patch('verifier.tasks.render_image', autospec=True) as mock_render_image, \
+            mock.patch('verifier.tasks.delete_source_files', autospec=True) as mock_delete_source_files, \
+            mock.patch('verifier.tasks.try_to_upload_blender_output_file', autospec=True) as mock_try_to_upload_file, \
+            mock.patch('verifier.tasks.load_images', autospec=True, return_value=self.loaded_images) as mock_load_images, \
+            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True) as mock_are_image_sizes_and_color_channels_equal, \
+            mock.patch('verifier.tasks.compare_images', side_effect=self._compare_images_negative) as mock_compare_images:  # noqa: E125
 
-        self.assertEqual(mock_clean_directory.call_count, 1)
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_upload_file_to_storage_cluster.call_count, 1)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 4)
-        self.assertEqual(mock_delete_file.call_count, 2)
-        mock_compare_ssim.assert_called_once()
-        mock_import_cv2.assert_called_once()
-        mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
+        self.assertEqual(mock_validate_downloaded_archives.call_count, 1)
+        self.assertEqual(mock_unpack_archives.call_count, 1)
+        mock_render_image.assert_called_once_with(1, self.output_format, self.scene_file, self.subtask_id)
+        self.assertEqual(mock_delete_source_files.call_count, 1)
+        self.assertEqual(mock_try_to_upload_file.call_count, 1)
+        self.assertEqual(mock_load_images.call_count, 1)
+        mock_are_image_sizes_and_color_channels_equal.assert_called_once_with(*self.loaded_images)
+        mock_compare_images.assert_called_once_with(self.mock_image1, self.mock_image2, self.subtask_id)
+        self.mock_verification_result.assert_called_once_with(
+            self.subtask_id,
             VerificationResult.MISMATCH.name,
         )
 
     def test_that_blender_verification_order_should_call_verification_result_with_result_error_if_download_fails(self):
         with mock.patch(
                 'verifier.tasks.download_archives_from_storage',
+                autospec=True,
                 side_effect=VerificationError(
                     'error',
                     ErrorCode.VERIFIER_FILE_DOWNLOAD_FAILED,
@@ -165,21 +143,10 @@ class VerifierVerificationIntegrationTest(ConcentIntegrationTestCase):
                 )
             ) as mock_download_archives_from_storage, \
             mock.patch('core.tasks.verification_result.delay', autospec=True) as mock_verification_result:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
 
-        mock_download_archives_from_storage.assert_called_once()
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
         mock_verification_result.assert_called_once_with(
             self.compute_task_def['subtask_id'],
             VerificationResult.ERROR.name,
@@ -187,312 +154,168 @@ class VerifierVerificationIntegrationTest(ConcentIntegrationTestCase):
             ErrorCode.VERIFIER_FILE_DOWNLOAD_FAILED.name,
         )
 
-    def test_blender_verification_order_should_call_verification_result_with_result_error_if_unpacking_archive_fails(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks, \
-            mock.patch('verifier.tasks.unpack_archive', side_effect=BadZipFile, autospec=True), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', return_value=['file_name']) as mock_get_files_list_from_archive, \
+    def test_that_blender_verification_order_should_call_verification_result_with_result_error_if_validation_of_downloaded_archives_fails(self):
+        with mock.patch('verifier.tasks.download_archives_from_storage', autospec=True) as mock_download_archives_from_storage, \
+            mock.patch(
+                'verifier.tasks.validate_downloaded_archives',
+                side_effect=VerificationError(
+                    'error',
+                    ErrorCode.VERIFIER_UNPACKING_ARCHIVE_FAILED,
+                    self.subtask_id
+                )
+            ),  \
             mock.patch('core.tasks.verification_result.delay', autospec=True) as mock_verification_result:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
 
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 2)
-        mock_clean_directory.assert_called_once_with(settings.VERIFIER_STORAGE_PATH)
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
         mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
+            self.subtask_id,
             VerificationResult.ERROR.name,
-            '',
+            'error',
             ErrorCode.VERIFIER_UNPACKING_ARCHIVE_FAILED.name,
         )
 
-    def test_blender_verification_order_should_call_verification_result_with_result_error_if_running_subprocess_raise_exception(self):
-        with mock.patch('verifier.tasks.clean_directory') as mock_clean_directory, \
-            mock.patch('verifier.tasks.send_request_to_storage_cluster') as mock_send_request_to_storage_cluster, \
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks') as mock_store_file_from_response_in_chunks, \
-            mock.patch('verifier.tasks.unpack_archive') as mock_unpack_archive, \
-            mock.patch('core.tasks.verification_result.delay', autospec=True) as mock_verification_result, \
-            mock.patch('verifier.tasks.get_files_list_from_archive', return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.run_blender', autospec=True, side_effect=SubprocessError('error')):  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
-
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 2)
-        mock_clean_directory.assert_called_once()
-        mock_send_request_to_storage_cluster.assert_called()
-        mock_store_file_from_response_in_chunks.assert_called()
-        mock_unpack_archive.assert_called()
-        mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
-            VerificationResult.ERROR.name,
-            'error',
-            ErrorCode.VERIFIER_RUNNING_BLENDER_FAILED.name,
-        )
-
-    def test_blender_verification_order_should_call_verification_result_with_result_error_if_running_subprocess_return_non_zero_code(self):
-        with mock.patch('verifier.tasks.clean_directory') as mock_clean_directory, \
-            mock.patch('verifier.tasks.send_request_to_storage_cluster') as mock_send_request_to_storage_cluster, \
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks') as mock_store_file_from_response_in_chunks, \
-            mock.patch('verifier.tasks.unpack_archive') as mock_unpack_archive, \
-            mock.patch('verifier.tasks.run_blender', mock_run_blender_with_error), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', return_value=['file_name']) as mock_get_files_list_from_archive, \
+    def test_blender_verification_order_should_call_verification_result_with_result_error_if_unpacking_archive_fails(self):
+        with  mock.patch('verifier.tasks.download_archives_from_storage', autospec=True) as mock_download_archives_from_storage, \
+            mock.patch('verifier.tasks.validate_downloaded_archives', autospec=True) as mock_validate_downloaded_archives, \
+            mock.patch('verifier.tasks.unpack_archives', side_effect=VerificationError(
+            "error",
+            ErrorCode.VERIFIER_UNPACKING_ARCHIVE_FAILED,
+            self.subtask_id
+        ), autospec=True), \
             mock.patch('core.tasks.verification_result.delay', autospec=True) as mock_verification_result:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
 
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 2)
-        mock_clean_directory.assert_called_once()
-        mock_send_request_to_storage_cluster.assert_called()
-        mock_store_file_from_response_in_chunks.assert_called()
-        mock_unpack_archive.assert_called()
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
+        self.assertEqual(mock_validate_downloaded_archives.call_count, 1)
         mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
+            self.subtask_id,
+            VerificationResult.ERROR.name,
+            'error',
+            ErrorCode.VERIFIER_UNPACKING_ARCHIVE_FAILED.name,
+        )
+
+    def test_blender_verification_order_should_call_verification_result_with_result_error_if_render_image_raises_exception(self):
+        with mock.patch('verifier.tasks.download_archives_from_storage', autospec=True) as mock_download_archives_from_storage, \
+            mock.patch('verifier.tasks.validate_downloaded_archives', autospec=True) as mock_validate_downloaded_archives, \
+            mock.patch('verifier.tasks.unpack_archives', autospec=True) as mock_unpack_archives, \
+            mock.patch('core.tasks.verification_result.delay', autospec=True) as mock_verification_result, \
+            mock.patch(
+                'verifier.tasks.render_image',
+                side_effect=VerificationError(
+                    'error',
+                    ErrorCode.VERIFIER_RUNNING_BLENDER_FAILED,
+                    self.subtask_id
+                ),
+                autospec=True):  # noqa: E125
+
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
+        self.assertEqual(mock_validate_downloaded_archives.call_count, 1)
+        self.assertEqual(mock_unpack_archives.call_count, 1)
+        mock_verification_result.assert_called_once_with(
+            self.subtask_id,
             VerificationResult.ERROR.name,
             'error',
             ErrorCode.VERIFIER_RUNNING_BLENDER_FAILED.name,
         )
 
-    def test_blender_verification_order_should_call_verification_result_with_result_error_if_opening_first_file_raise_memory_error(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks,\
-            mock.patch('verifier.tasks.unpack_archive', autospec=True) as mock_unpack_archive,\
-            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result,\
-            mock.patch('verifier.tasks.run_blender', mock_run_blender), \
-            mock.patch('builtins.open', autospec=True, side_effect=MemoryError('error')), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.delete_file') as mock_delete_file:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
+    def test_blender_verification_order_should_call_verification_result_with_result_error_if_load_images_fails(self):
+        with mock.patch('verifier.tasks.download_archives_from_storage', autospec=True) as mock_download_archives_from_storage, \
+            mock.patch('verifier.tasks.validate_downloaded_archives', autospec=True) as mock_validate_downloaded_archives, \
+            mock.patch('verifier.tasks.unpack_archives', autospec=True) as mock_unpack_archives, \
+            mock.patch('verifier.tasks.render_image', autospec=True) as mock_render_image, \
+            mock.patch('verifier.tasks.delete_source_files', autospec=True) as mock_delete_source_files, \
+            mock.patch('verifier.tasks.try_to_upload_blender_output_file', autospec=True) as mock_try_to_upload_file, \
+            mock.patch(
+                'verifier.tasks.load_images',
+                autospec=True,
+                side_effect=VerificationError(
+                    'error',
+                    ErrorCode.VERIFIER_LOADING_FILES_INTO_MEMORY_FAILED,
+                    self.subtask_id
+                ),
+            ), \
+            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result:  # noqa: E125
 
-        self.assertEqual(mock_clean_directory.call_count, 1)
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 3)
-        self.assertEqual(mock_delete_file.call_count, 2)
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
+        self.assertEqual(mock_validate_downloaded_archives.call_count, 1)
+        self.assertEqual(mock_unpack_archives.call_count, 1)
+        mock_render_image.assert_called_once_with(1, self.output_format, self.scene_file, self.subtask_id)
+        self.assertEqual(mock_delete_source_files.call_count, 1)
+        self.assertEqual(mock_try_to_upload_file.call_count, 1)
         mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
+            self.subtask_id,
             VerificationResult.ERROR.name,
             'error',
             ErrorCode.VERIFIER_LOADING_FILES_INTO_MEMORY_FAILED.name,
-        )
-
-    def test_blender_verification_order_should_continue_verification_if_opening_blender_output_file_raises_os_error(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks,\
-            mock.patch('verifier.tasks.unpack_archive', autospec=True) as mock_unpack_archive,\
-            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result,\
-            mock.patch('verifier.tasks.run_blender', mock_run_blender), \
-            mock.patch('builtins.open', autospec=True, side_effect=OSError('error')), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.compare_ssim', return_value=1.0) as mock_compare_ssim, \
-            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True), \
-            mock.patch('verifier.tasks.delete_file') as mock_delete_file, \
-            mock.patch('verifier.tasks.import_cv2', return_value=self._prepare_cv2_mock()) as mock_import_cv2:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
-
-        self.assertEqual(mock_clean_directory.call_count, 1)
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 4)
-        self.assertEqual(mock_delete_file.call_count, 2)
-        mock_compare_ssim.assert_called_once()
-        mock_import_cv2.assert_called_once()
-        mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
-            VerificationResult.MATCH.name,
-        )
-
-    def test_blender_verification_order_should_call_verification_result_with_result_error_if_cv2_imread_raise_memory_error(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks,\
-            mock.patch('verifier.tasks.unpack_archive', autospec=True) as mock_unpack_archive, \
-            mock.patch('verifier.tasks.upload_file_to_storage_cluster', autospec=True) as mock_upload_file_to_storage_cluster, \
-            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result,\
-            mock.patch('verifier.tasks.run_blender', mock_run_blender), \
-            mock.patch('builtins.open', autospec=True, side_effect=[io.BytesIO(b'test'), io.BytesIO(b'test')]), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', autospec=True, return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.delete_file') as mock_delete_file, \
-            mock.patch('verifier.tasks.import_cv2', return_value=self._prepare_cv2_mock('exception')) as mock_import_cv2:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
-
-        self.assertEqual(mock_clean_directory.call_count, 1)
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_upload_file_to_storage_cluster.call_count, 1)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 4)
-        self.assertEqual(mock_delete_file.call_count, 2)
-        mock_import_cv2.assert_called_once()
-        mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
-            VerificationResult.ERROR.name,
-            'error',
-            ErrorCode.VERIFIER_LOADING_FILES_INTO_MEMORY_FAILED.name,
-        )
-
-    def test_blender_verification_order_should_call_verification_result_with_result_error_if_cv2_imread_returns_none(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks,\
-            mock.patch('verifier.tasks.unpack_archive', autospec=True) as mock_unpack_archive, \
-            mock.patch('verifier.tasks.upload_file_to_storage_cluster', autospec=True) as mock_upload_file_to_storage_cluster, \
-            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result,\
-            mock.patch('verifier.tasks.run_blender', mock_run_blender), \
-            mock.patch('builtins.open', autospec=True, side_effect=[io.BytesIO(b'test'), io.BytesIO(b'test')]), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', autospec=True, return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True), \
-            mock.patch('verifier.tasks.delete_file') as mock_delete_file, \
-            mock.patch('verifier.tasks.import_cv2', return_value=self._prepare_cv2_mock('none')) as mock_import_cv2:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
-
-        self.assertEqual(mock_clean_directory.call_count, 1)
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_upload_file_to_storage_cluster.call_count, 1)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 4)
-        self.assertEqual(mock_delete_file.call_count, 2)
-        mock_import_cv2.assert_called_once()
-        mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
-            VerificationResult.ERROR.name,
-            'Loading files using OpenCV fails.',
-            ErrorCode.VERIFIER_LOADING_FILES_WITH_OPENCV_FAILED.name,
         )
 
     def test_blender_verification_order_should_call_verification_result_with_result_error_if_compare_ssim_raise_value_error(self):
-        with mock.patch('verifier.tasks.clean_directory', autospec=True) as mock_clean_directory,\
-            mock.patch('verifier.tasks.send_request_to_storage_cluster', autospec=True) as mock_send_request_to_storage_cluster,\
-            mock.patch('verifier.tasks.store_file_from_response_in_chunks', autospec=True) as mock_store_file_from_response_in_chunks,\
-            mock.patch('verifier.tasks.unpack_archive', autospec=True) as mock_unpack_archive,\
-            mock.patch('verifier.tasks.upload_file_to_storage_cluster', autospec=True) as mock_upload_file_to_storage_cluster,\
-            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result,\
-            mock.patch('verifier.tasks.run_blender', mock_run_blender), \
-            mock.patch('builtins.open', autospec=True, side_effect=[io.BytesIO(b'test'), io.BytesIO(b'test')]), \
-            mock.patch('verifier.tasks.get_files_list_from_archive', return_value=['file_name']) as mock_get_files_list_from_archive, \
-            mock.patch('verifier.tasks.compare_ssim', side_effect=ValueError('error')) as mock_compare_ssim, \
-            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True), \
-            mock.patch('verifier.tasks.delete_file', autospec=True) as mock_delete_file, \
-            mock.patch('verifier.tasks.import_cv2', return_value=self._prepare_cv2_mock()) as mock_import_cv2:  # noqa: E125
-            blender_verification_order(
-                subtask_id=self.compute_task_def['subtask_id'],
-                source_package_path=self.source_package_path,
-                source_size=self.report_computed_task.task_to_compute.size,
-                source_package_hash=self.report_computed_task.task_to_compute.package_hash,
-                result_package_path=self.result_package_path,
-                result_size=self.report_computed_task.size,  # pylint: disable=no-member
-                result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
-                output_format=BlenderSubtaskDefinition.OutputFormat[
-                    self.compute_task_def['extra_data']['output_format']
-                ].name,
-                scene_file=self.compute_task_def['extra_data']['scene_file'],
-            )
+        with mock.patch('verifier.tasks.download_archives_from_storage', autospec=True) as mock_download_archives_from_storage, \
+            mock.patch('verifier.tasks.validate_downloaded_archives', autospec=True) as mock_validate_downloaded_archives, \
+            mock.patch('verifier.tasks.unpack_archives', autospec=True) as mock_unpack_archives, \
+            mock.patch('verifier.tasks.render_image', autospec=True) as mock_render_image, \
+            mock.patch('verifier.tasks.delete_source_files', autospec=True) as mock_delete_source_files, \
+            mock.patch('verifier.tasks.try_to_upload_blender_output_file', autospec=True) as mock_try_to_upload_file, \
+            mock.patch('verifier.tasks.load_images', autospec=True, return_value=self.loaded_images) as mock_load_images, \
+            mock.patch('verifier.tasks.are_image_sizes_and_color_channels_equal', return_value=True) as mock_are_image_sizes_and_color_channels_equal, \
+            mock.patch(
+                'verifier.tasks.compare_images',
+                autospec=True,
+                side_effect=VerificationError(
+                    'error',
+                    ErrorCode.VERIFIER_COMPUTING_SSIM_FAILED,
+                    self.subtask_id
+                ),
+            ), \
+            mock.patch('verifier.tasks.verification_result.delay', autospec=True) as mock_verification_result:  # noqa: E125
 
-        self.assertEqual(mock_clean_directory.call_count, 1)
-        self.assertEqual(mock_send_request_to_storage_cluster.call_count, 2)
-        self.assertEqual(mock_store_file_from_response_in_chunks.call_count, 2)
-        self.assertEqual(mock_unpack_archive.call_count, 2)
-        self.assertEqual(mock_upload_file_to_storage_cluster.call_count, 1)
-        self.assertEqual(mock_get_files_list_from_archive.call_count, 4)
-        self.assertEqual(mock_delete_file.call_count, 2)
-        mock_compare_ssim.assert_called_once()
-        mock_import_cv2.assert_called_once()
+            self._send_blender_verification_order()
+
+        self.assertEqual(mock_download_archives_from_storage.call_count, 1)
+        self.assertEqual(mock_validate_downloaded_archives.call_count, 1)
+        self.assertEqual(mock_unpack_archives.call_count, 1)
+        mock_render_image.assert_called_once_with(1, self.output_format, self.scene_file, self.subtask_id)
+        self.assertEqual(mock_delete_source_files.call_count, 1)
+        self.assertEqual(mock_try_to_upload_file.call_count, 1)
+        self.assertEqual(mock_load_images.call_count, 1)
+        mock_are_image_sizes_and_color_channels_equal.assert_called_once_with(*self.loaded_images)
         mock_verification_result.assert_called_once_with(
-            self.compute_task_def['subtask_id'],
+            self.subtask_id,
             VerificationResult.ERROR.name,
             'error',
             ErrorCode.VERIFIER_COMPUTING_SSIM_FAILED.name,
+        )
+
+    def _compare_images_positive(self, _image_1, _image_2, subtask_id):
+        self.mock_verification_result(
+            subtask_id,
+            VerificationResult.MATCH.name
+        )
+
+    def _compare_images_negative(self, _image_1, _image_2, subtask_id):
+        self.mock_verification_result(
+            subtask_id,
+            VerificationResult.MISMATCH.name
+        )
+
+    def _send_blender_verification_order(self):
+        blender_verification_order(
+            subtask_id=self.subtask_id,
+            source_package_path=self.source_package_path,
+            source_size=self.report_computed_task.task_to_compute.size,
+            source_package_hash=self.report_computed_task.task_to_compute.package_hash,
+            result_package_path=self.result_package_path,
+            result_size=self.report_computed_task.size,  # pylint: disable=no-member
+            result_package_hash=self.report_computed_task.package_hash,  # pylint: disable=no-member  # pylint: disable=no-member
+            output_format=BlenderSubtaskDefinition.OutputFormat[
+                self.output_format
+            ].name,
+            scene_file=self.scene_file,
         )
